% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{runSircleStats}
\alias{runSircleStats}
\title{runSircleStats}
\usage{
runSircleStats(sv, condLabel, cond0, cond1)
}
\arguments{
\item{sv}{a pretrained instance of the stats model}

\item{condLabel}{The condition label (e.g. column) that exists in your patient file and the sample files (e.g. gender)}

\item{cond0}{Condition 0 (i.e, female)}

\item{cond1}{e.g. your condition 1 value (e.g. male)}
}
\value{
df: a DF with the stats in it
}
\description{
Runs stats on two groups of patients by integrating their data and computing stats on the latent dimensions.
}
