% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCMVis.R
\name{sircleAlluvial_MRP}
\alias{sircleAlluvial_MRP}
\title{sircleAlluvial}
\usage{
sircleAlluvial_MRP(filename, outputFilename = "SiRCle_Alluvial.pdf")
}
\arguments{
\item{filename}{Path to the input file}
}
\value{
alluvialPlt The plot generated
}
\description{
This function creates an alluvial plot from an input CSV to create an alluvial plot for the RCM_MRP workflow.
}
