% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleGSEA}
\alias{sircleGSEA}
\title{sircleGSEA}
\usage{
sircleGSEA(
  rcm,
  df,
  geneId,
  numNodes,
  userPathways,
  fgseaPCutoff = 1,
  fileLabel = "GSEA",
  groupLabels = c("MDE", "MDS", "TMDE", "TMDS", "TPDE", "TPDE_TMDS", "TPDS",
    "MDS_TMDE", "TPDS_TMDE", "MDE-ncRNA", "MDS-ncRNA")
)
}
\arguments{
\item{rcm}{Instance of RCM after running sircleRCM}

\item{df}{Dataframe with the results from the sircleRCM}

\item{geneId}{Gene ID column name in you dataframe}

\item{numNodes}{Number of latent nodes in your VAE}

\item{fgseaPCutoff}{\emph{Optional: } Padjusted cutoff for GSEA data (the other results aren't saved) \strong{Default=0.2}}

\item{groupLabels}{\emph{Optional: } Labels from the groups  \strong{Default=c('MDE', 'MDS', 'TMDE', 'TMDS', 'TPDE', 'TPDE_TMDS', 'TPDS', 'MDS_TMDE', 'TPDS_TMDE', "MDE-ncRNA", "MDS-ncRNA")}}

\item{pathway_list_go}{A list of pathways for the GSEA (see example for details)}
}
\description{
Uses scircm to compute the regulatory clustering model.
}
