% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTF.R
\name{sircleMotifFimo}
\alias{sircleMotifFimo}
\title{sircleMotifFimo}
\usage{
sircleMotifFimo(
  geneId,
  fimoFilename,
  sircleRCMFilename,
  TFPadjCol,
  TFValueCol,
  pDisplayCutoff = 0.05,
  TFInDataset = T,
  fimoPcol = "p-value",
  outputDir = "",
  regLabels = "RegulatoryLabels",
  outputFilename = "SiRCle-motif.csv"
)
}
\arguments{
\item{geneId}{Column name with the gene identifier}

\item{fimoFilename}{Filename/path of the output from fimo (usually fimo.tsv)}

\item{sircleRCMFilename}{Filename/path of the output from sircleRCM}

\item{TFPadjCol}{Column name of the padjusted value in the sircleRCMFilename file (i.e. your RNAseq, or your protein p column)}

\item{TFValueCol}{Column name of  rna value for the TF (could also use the proteomics logfc)}

\item{pDisplayCutoff}{Cutoff for the clusters (only returns results less than this, usually 0.05)}

\item{TFInDataset}{Whether or not you require the TF to be expressed in the dataframe}

\item{fimoPcol}{Column name of p or q value used to filter fimo results(q-value for adjusted or p-value for un-adj)}

\item{regLabels}{Column name of the regulatory labels}

\item{outputFilename}{Filename for the output file.}
}
\value{
scimo an instance of the scimo package
}
\description{
Uses FIMO to predict TF binding and then finds enrichment in clusters.
}
