% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleORA.R
\name{sircleORA}
\alias{sircleORA}
\title{sircleORA}
\usage{
sircleORA(
  InputData,
  geneID,
  regLabels = "RegulatoryLabels",
  PriorKnowledge,
  PKName = "",
  emptyRegLabel = "",
  RemoveBackgroundGenes = "TRUE",
  Plot_FileType = "pdf",
  minGSSize = 10,
  maxGSSize = 1000,
  Plot_p.adj = 0.2,
  Plot_Percentage = 10,
  OutputFileName = ""
)
}
\arguments{
\item{InputData}{DF with column gene names amd regulatory labels (i.e. sircle clusters). Usually output from SiRCle RCM functions.}

\item{geneID}{Provide column name for the gene ID in InputData. Needs to match the the gene ID in the pathway file provided}

\item{regLabels}{\emph{Optional: } Column name of the column with the regulatory labels in InputData. \strong{default: "RegulatoryLabels"}}

\item{PriorKnowledge}{DF including the prior knowledge (PK), ie. pathway file. File must include column "term" with the pathway or TF-regulon name, column "gene" with the gene name and column "Description" with description that will be depicted on the plots.}

\item{PKName}{\emph{Optional: } Name of the prior knowledge used, i.e. pathway name such as KEGG \strong{default: ""}}

\item{emptyRegLabel}{\emph{Optional: } Column name of the empty regulatory group \strong{default: ""}}

\item{Plot_FileType}{\emph{Optional: } Output file type for the figures one of: "svg", "pdf", "png" \strong{default: "pdf"}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{maxGSSize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{Plot_p.adj}{\emph{Optional: } q value cutoff from ORA that should be plotted \strong{default: 0.2}}

\item{Plot_Percentage}{\emph{Optional: } Percentage of genes that are detected of a pathway \strong{default: 10}}

\item{OutputFileName}{\emph{Optional: } Output filename \strong{default: ""}}

\item{RemoveBackgroundGenes\emph{Optional:}{} If TRUE, genes that fall into background based on the choosen Background method for SiRCle RCM are removed from the universe. \strong{default: "TRUE"}}
}
\value{
DF with the results of the over representation analysis
}
\description{
Uses enricher to run ORA on each of the clusters as defined by the regulatory labels using prior knowledge such as gene-sets
}
