% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleORA.R
\name{sircleORAHuman_Enrich}
\alias{sircleORAHuman_Enrich}
\title{sircleORAHuman_Enrich}
\usage{
sircleORAHuman_Enrich(
  filename,
  regLabels = "RegulatoryLabels",
  emptyRegLabel = "",
  RemoveBackgroundGenes = "TRUE",
  enricher_geneID,
  enricher_Pathways,
  enricher_PathwayName = "",
  fileType = "pdf",
  minGSSize = 10,
  maxGSSize = 1000,
  Plot_p.adj = 0.2,
  Plot_Percentage = 10,
  outputFolder = ""
)
}
\arguments{
\item{filename}{Path to the input file}

\item{regLabels}{\emph{Optional: } regLabels The label of the column with the regulatory labels \strong{default: "RegulatoryLabels"}}

\item{emptyRegLabel}{\emph{Optional: } emptyRegLabel The label of the empty regulatory group \strong{default: ""}}

\item{enricher_geneID}{Provide column name for the gene ID. Needs to match the the gene ID in the pathway file provided}

\item{enricher_Pathways}{Provide pathway file. Pathway file must include column "term" with the pathway name, column "gene" with the gene name and column "Description" with pathway description that will be depicted on the plots.}

\item{enricher_PathwayName}{\emph{Optional: } Name of the pathway list used \strong{default: ""}}

\item{fileType}{\emph{Optional: } fileType Output file type for the figures one of: "svg", "pdf", "png" \strong{default: "pdf"}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{maxGSSize}{\emph{Optional: } maximum group size in ORA \strong{default: 1000}}

\item{Plot_p.adj}{\emph{Optional: } q value cutoff from ORA that should be plotted \strong{default: 0.2}}

\item{Plot_Percentage}{\emph{Optional: } Percentage of genes that are detected of a pathway \strong{default: 10}}
}
\description{
Uses enricher to run ORA on each of the clusters as defined by the regulatory labels using a selfdefined pathway list (on HUMAN)
}
