% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sirclePlot}
\alias{sirclePlot}
\title{sirclePlot}
\usage{
sirclePlot(filename, regLabels = "RegulatoryLabels", fileType = "pdf")
}
\arguments{
\item{filename}{filename of the RCM}

\item{regLabels}{Labels of the regulatory file}

\item{figType}{\emph{Optional: } File ending for the saved figures \strong{Default="pdf"}}
}
\description{
Uses scircm to and plots a circle plot of the sircle regulatory groups.
}
