% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleRCM}
\alias{sircleRCM}
\title{sircleRCM}
\usage{
sircleRCM(
  rnaFile,
  methFile,
  proteinFile,
  geneId,
  rnaValueCol,
  rnaPadjCol,
  methValueCol,
  methPadjCol,
  proteinValueCol,
  proteinPadjCol,
  proteinCols = NULL,
  rnaPadjCutoff = 0.05,
  rnaLogFCCutoff = 0.5,
  proteinPadjCutoff = 0.05,
  proteinValueCutoff = 0.3,
  methPadjCutoff = 0.05,
  methDiffCutoff = 10,
  backgroundMethod = "P|(M&R)",
  fileSep = ",",
  nonCodingGeneList = NULL,
  outputFileName = "SiRCle_RCM.csv",
  logfile = "logfileRCM.csv",
  envName = NULL,
  condaEnvName = NULL,
  envPath = NULL
)
}
\arguments{
\item{rnaFile}{Filename for your RNAseq data (results from DeSeq2 and also your normalised expression counts)}

\item{methFile}{Filename for your DNA methylation data (results from differential methylation analysis)}

\item{proteinFile}{Filename/path of you Protein data (results from DeSeq2 and also your normalised expression counts)}

\item{geneId}{Column name of geneId this MUST BE THE SAME in each of your protein, RNAseq and DNAmethylation files (we join on this)}

\item{rnaValueCol}{Column name of RNA value in rnaFile (usually logFoldChange_r)MUST BE UNIQUE BETWEEN methFile and proteinFile}

\item{rnaPadjCol}{Column name of RNA p adjusted value in rnaFile (usually padj_r) MUST BE UNIQUE BETWEEN methFile and proteinFile}

\item{methValueCol}{Column name of Methylation difference value in methFile (usually meth.diff) MUST BE UNIQUE BETWEEN rnaFile and proteinFile}

\item{methPadjCol}{Column name of Methylation p adjusted value in methFile (usually padj_m) MUST BE UNIQUE BETWEEN rnaFile and proteinFile}

\item{proteinValueCol}{Column name of protein log fold change in proteinFile (usually logFC_p) MUST BE UNIQUE BETWEEN rnaFile and methFile}

\item{proteinPadjCol}{Column name of protein p adjusted value in proteinFile (usually padh_p) MUST BE UNIQUE BETWEEN rnaFile and proteinFile}

\item{proteinCols}{A list of columns that you want to impute missing values for (optional) jsut uses minimum value (for VAE)}

\item{rnaPadjCutoff}{\emph{Optional: }Padjusted cutoff for RNAseq data \strong{Default=0.05}}

\item{rnaLogFCCutoff}{\emph{Optional: } LogFoldchange cutoff for RNAseq data \strong{Default=0.5}}

\item{proteinPadjCutoff}{\emph{Optional: } Padjusted cutoff for Protein data \strong{Default=0.05}}

\item{proteinValueCutoff}{\emph{Optional: } LogFoldchange cutoff for Protein data \strong{Default=0.3}}

\item{methPadjCutoff}{\emph{Optional: } Padjusted cutoff for DNA methylation data \strong{Default=0.05}}

\item{methDiffCutoff}{\emph{Optional: } DNA Methylation difference cutoff for DNA methylation \strong{Default=10}}

\item{backgroundMethod}{\emph{Optional: } Background method (NEED Description for each one here) \strong{Default="P|(M&R)"}}

\item{fileSep}{\emph{Optional: } Separator for files i.e. expecting CSV's however if they are all TSVs you can change this \strong{Default=","}}

\item{nonCodingGeneList}{\emph{Optional: } List of genes that are annotated to be non-coding related \strong{Default=NULL}}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default=SiRCle_RCM.csv}}

\item{logfile}{\emph{Optional: } Name of the logfile \strong{Default=logfileRCM.csv}}

\item{envName}{\emph{Optional: } Name of your previously setup python virtual environment \strong{Default=NULL}}

\item{condaEnvName}{\emph{Optional: } Name of your previously setup python conda environment \strong{Default=NULL}}

\item{envPath}{\emph{Optional: } Path as a string to your previously setup python \strong{Default=NULL}}
}
\value{
rcm an instance of the rcm package
}
\description{
Uses scircm to compute the regulatory clustering model.
}
