% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleRCM_2Cond}
\alias{sircleRCM_2Cond}
\title{sircleRCM_2Cond}
\usage{
sircleRCM_2Cond(
  Cond1_File,
  Cond2_File,
  geneID,
  Cond1ValueCol = "Log2FC",
  Cond1PadjCol = "padj",
  Cond2ValueCol = "Log2FC",
  Cond2PadjCol = "padj",
  Cond1_padj_cutoff = 0.05,
  Cond2_padj_cutoff = 0.05,
  Cond1_FC_cutoff = 1,
  Cond2_FC_cutoff = 0.5,
  backgroundMethod = "C1&C2",
  OutputFileName = "Sircle_RCM.csv"
)
}
\arguments{
\item{geneID}{Column name of geneId this MUST BE THE SAME in each of your Input files (we join on this)}

\item{Cond1ValueCol}{Column name of log fold change in Cond1File}

\item{Cond1PadjCol}{Column name of p adjusted value in Cond1File}

\item{Cond2ValueCol}{Column name of log fold change in Cond2File}

\item{Cond2PadjCol}{Column name of protein p adjusted value in Cond2File}

\item{backgroundMethod}{\emph{Optional: } Background method C1|C2, C1&C2, C2, C1 or * \strong{Default="C1&C2"}}

\item{Cond1File}{DF for your data (results from e.g. DeSeq2)}

\item{Cond2File}{DF for your data (results from e.g. DeSeq2)}

\item{Cond1PadjCutoff}{\emph{Optional: }Padjusted cutoff forCond1 data \strong{Default=0.05}}

\item{Cond1LogFCCutoff}{\emph{Optional: } LogFoldchange cutoff for Cond1 data \strong{Default=0.5}}

\item{Cond2FilePadjCutoff}{\emph{Optional: } Padjusted cutoff for Cond2File data \strong{Default=0.05}}

\item{Cond2FileCutoff}{\emph{Optional: } LogFoldchange cutoff for Cond2File data \strong{Default=0.5}}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default=SiRCle_RCM.csv}}
}
\value{
rcm an instance of the rcm package
}
\description{
Computes the regulatory clustering model (RCM) using the SiRCle regulatory rules between two conditions of the same data type.
}
