% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleRCM_MRP}
\alias{sircleRCM_MRP}
\title{sircleRCM_MRP}
\usage{
sircleRCM_MRP(
  methFile,
  rnaFile,
  protFile,
  geneID,
  rnaValueCol = "Log2FC",
  rnaPadjCol = "padj",
  methValueCol = "Diff",
  methPadjCol = "padj",
  proteinValueCol = "Log2FC",
  proteinPadjCol = "padj",
  rna_padj_cutoff = 0.05,
  prot_padj_cutoff = 0.05,
  meth_padj_cutoff = 0.05,
  rna_FC_cutoff = 1,
  prot_FC_cutoff = 0.5,
  meth_Diff_cutoff = 0.1,
  backgroundMethod,
  OutputFileName = "Sircle_RCM"
)
}
\arguments{
\item{methFile}{DF for your DNA methylation data (results from differential methylation analysis)}

\item{rnaFile}{DF for your RNAseq data (results from DeSeq2 and also your normalised expression counts)}

\item{geneID}{Column name of geneId this MUST BE THE SAME in each of your protein, RNAseq and DNAmethylation files (we join on this)}

\item{rnaValueCol}{Column name of RNA value in rnaFile}

\item{rnaPadjCol}{Column name of RNA p adjusted value in rnaFile}

\item{methValueCol}{Column name of Methylation difference value in methFile}

\item{methPadjCol}{Column name of Methylation p adjusted value in methFile}

\item{proteinValueCol}{Column name of protein log fold change in proteinFile}

\item{proteinPadjCol}{Column name of protein p adjusted value in proteinFile}

\item{backgroundMethod}{\emph{Optional: } Background methods: "P|(M&R)", "P|M|R", "P|R", "P&R", "P&M&R", "(P&M)|(P&R)|(M&R)", "(P&M)|(P&R)" or "*". For details on the methods, please check https://github.com/ArianeMora/SiRCleR/blob/main/vignettes/SiRCle_RCM_Notebook.Rmd \strong{Default="P|(M&R)"}}

\item{proteinFile}{DF of you Protein data (results from DeSeq2 and also your normalised expression counts)}

\item{rnaPadjCutoff}{\emph{Optional: }Padjusted cutoff for RNAseq data \strong{Default=0.05}}

\item{rnaLogFCCutoff}{\emph{Optional: } LogFoldchange cutoff for RNAseq data \strong{Default=0.5}}

\item{proteinPadjCutoff}{\emph{Optional: } Padjusted cutoff for Protein data \strong{Default=0.05}}

\item{proteinValueCutoff}{\emph{Optional: } LogFoldchange cutoff for Protein data \strong{Default=0.3}}

\item{methPadjCutoff}{\emph{Optional: } Padjusted cutoff for DNA methylation data \strong{Default=0.05}}

\item{methDiffCutoff}{\emph{Optional: } DNA Methylation difference cutoff for DNA methylation \strong{Default=10}}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default="SiRCle_RCM"}}
}
\value{
DF with input data combined by geneID and SiRCle Clusters
}
\description{
Computes the regulatory clustering model (RCM) using the SiRCle regulatory rules for DNA-methylation, RNAseq and Proteomics data layers (MRP).
}
