% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleRCM.R
\name{sircleRCM_RP}
\alias{sircleRCM_RP}
\title{sircleRCM_RP}
\usage{
sircleRCM_RP(
  rnaFile,
  protFile,
  geneID,
  rnaValueCol = "Log2FC",
  rnaPadjCol = "padj",
  proteinValueCol = "Log2FC",
  proteinPadjCol = "padj",
  rna_padj_cutoff = 0.05,
  prot_padj_cutoff = 0.05,
  rna_FC_cutoff = 1,
  prot_FC_cutoff = 0.5,
  backgroundMethod = "P&R",
  OutputFileName = "Sircle_RCM.csv"
)
}
\arguments{
\item{rnaFile}{Filename for your RNAseq data (results from DeSeq2 and also your normalised expression counts)}

\item{geneID}{Column name of geneId this MUST BE THE SAME in each of your protein, RNAseq and DNAmethylation files (we join on this)}

\item{rnaValueCol}{Column name of RNA value in rnaFile}

\item{rnaPadjCol}{Column name of RNA p adjusted value in rnaFile}

\item{proteinValueCol}{Column name of protein log fold change in proteinFile}

\item{proteinPadjCol}{Column name of protein p adjusted value in proteinFile}

\item{backgroundMethod}{\emph{Optional: } Background method (NEED Description for each one here) \strong{Default="P&R"}}

\item{proteinFile}{Filename/path of you Protein data (results from DeSeq2 and also your normalised expression counts)}

\item{rnaPadjCutoff}{\emph{Optional: }Padjusted cutoff for RNAseq data \strong{Default=0.05}}

\item{rnaLogFCCutoff}{\emph{Optional: } LogFoldchange cutoff for RNAseq data \strong{Default=0.5}}

\item{proteinPadjCutoff}{\emph{Optional: } Padjusted cutoff for Protein data \strong{Default=0.05}}

\item{proteinValueCutoff}{\emph{Optional: } LogFoldchange cutoff for Protein data \strong{Default=0.3}}

\item{outputFileName}{\emph{Optional: } Output filename \strong{Default=SiRCle_RCM.csv}}
}
\value{
rcm an instance of the rcm package
}
\description{
Computes the regulatory clustering model (RCM) using the SiRCle regulatory rules for RNAseq and Proteomics data layers (RP).
}
