% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTFVis.R
\name{sircleTFEman}
\alias{sircleTFEman}
\title{sircleTFEman}
\usage{
sircleTFEman(
  TfDf,
  colorLabel,
  fillNAVal = NA,
  titleStr = "",
  continuous = T,
  plotBar = T,
  plotEman = T
)
}
\arguments{
\item{TfDf}{Column name with the gene identifier}

\item{colorLabel}{Column with the value that we will filter on (usually the dorothea column)}

\item{fillNAVal}{\emph{Optional:}  Fill value for None columns \strong{Default: NA}}

\item{titleStr}{\emph{Optional:} String for the title \strong{Default: ""}}

\item{continuous}{\emph{Optional:} Whether or not the data are continous \strong{Default: T}}

\item{plotBar}{\emph{Optional:} Whether or not to plot the bar charts \strong{Default: T}}

\item{plotEman}{\emph{Optional:} Whether or not to plot the connection plot \strong{Default: T}}
}
\value{
scimo an instance of the scimo package
}
\description{
Creates the eman plot for the TF vis
}
