% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToyData.R
\name{ToyData}
\alias{ToyData}
\title{Access built-in example data}
\usage{
ToyData(Dataset)
}
\arguments{
\item{Dataset}{Character: name of a built-in dataset:
\itemize{
\item{\code{"DNA_Methylation"}: }
\item{\code{"RNAseq"}: }
\item{\code{"Proteomics"}: }
\item{\code{"Hallmarks"}: }
}}
}
\value{
A data frame containing the toy data.
}
\description{
Import and process .csv file to create toy data DF.
}
\examples{
Proteomics <- SiRCleR::ToyData("Proteomics")

}
