% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleTF.R
\name{sircleMotif}
\alias{sircleMotif}
\title{sircleMotif}
\usage{
sircleMotif(
  clusterGeneId,
  doroFile,
  sircleRCMFilename,
  TFPadjCol,
  TFValueCol,
  TargetPadjCol,
  TargetValueCol,
  RegulatoryLabel,
  TFInDataset = T,
  doroLevel = c("A"),
  clusters = c("MDS", "MDE", "TPDE", "TPDS"),
  outputFilename = "SiRCle-motif.csv",
  outputDir = "",
  plotOn = T
)
}
\arguments{
\item{clusterGeneId}{Column name with the gene identifier (note this must match the TF ids used in DoRoTHea i.e. gene name)}

\item{doroFile}{Filename/path of the output from Dorothea}

\item{sircleRCMFilename}{Filename/path of the output from sircleRCM}

\item{TFPadjCol}{Column name of the padjusted value in the sircleRCMFilename file (i.e. your RNAseq, or your protein p column)}

\item{TFValueCol}{Column name of  rna value for the TF (could also use the proteomics logfc)}

\item{TFInDataset}{Whether or not you require the TF to be expressed in the dataframe}

\item{doroLevel}{level(s) from Dorothea (i.e. A - D) see DoRoThea for more info.}

\item{clusters}{clusters to test for enrichment.}

\item{outputFilename}{Filename for the output file.}

\item{plotOn}{Whether or not to plot the results}

\item{pDisplayCutoff}{Cutoff for the clusters (only returns results less than this, usually 0.05)}

\item{ouputDir}{Directory for saving}
}
\value{
DF of the results
}
\description{
Uses known TF interactions from DoRoThea and then finds enrichment of TF-Target relationships in clusters.
}
