% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircleORA.R
\name{sircleORA}
\alias{sircleORA}
\title{sircleORA}
\usage{
sircleORA(
  filename,
  entrezId,
  regLabels = "RegulatoryLabels",
  emptyRegLabel = "",
  RemoveBackgroundGenes = "TRUE",
  fileType = "pdf",
  minGSSize = 10,
  qvalueCutoff = 0.2,
  pvalueCutoff = 0.05,
  showCatagory = 30,
  OutputFileName = "",
  Organism = "Human"
)
}
\arguments{
\item{filename}{Input file}

\item{entrezId}{Column name for the entrez ID}

\item{regLabels}{\emph{Optional: } regLabels The label of the column with the regulatory labels \strong{default: "RegulatoryLabels"}}

\item{emptyRegLabel}{\emph{Optional: } emptyRegLabel The label of the empty regulatory group \strong{default: ""}}

\item{fileType}{\emph{Optional: } fileType Output file type for the figures one of: "svg", "pdf", "png" \strong{default: "pdf"}}

\item{minGSSize}{\emph{Optional: } minimum group size in ORA \strong{default: 10}}

\item{qvalueCutoff}{\emph{Optional: } q value cutoff from ORA \strong{default: 0.2}}

\item{OutputFileName}{Output filename}

\item{Organism}{\emph{Optional: } Choose either "Human" or "Mouse" \strong{default: Human"}}

\item{RemoveBackgroundGenes\emph{Optional:}{} If TRUE, genes that fall into background based on the choosen Background method for SiRCle RCM are removed from the universe. \strong{default: "TRUE"}}
}
\value{
sircleORA an instance of the sircle package
}
\description{
Uses encrichGo to run ORA on each of the clusters as defined by the regulatory labels using GO terms pathways
}
